requirements.txt
PyEnchant
